Use django_overwrites app from partners
